fix: isolate agent runtime lock validation
